Revert changes to how users are compared in queries
The reverted code made the assumption that the User
always uses the Id as SecurityInfo.Identifier. The whole point
with the SecurityInfo is that you can choose any field as the Identifier.

The problems with second level cache found earlier by wut was not a problem
with the queries. The problem was that without Equals/GetHashCode on User
Nhibernate cannot compare objects when removing from collections. So nothing
was removed from collections and the tests failed.